pyproject.toml
[project]
name = "pylizlib"
version = "0.1.55"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psutil>=7.1.0",
    "python-dotenv>=1.1.0",
    "pywin32>=1.0 ; sys_platform == 'win32'",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "typer>=0.15.2",
]

[project.scripts]
pyliz = "pylizlib.cli:app"

[project.optional-dependencies]
qt = [
    "pyside6>=6.9.2",
]
media = [
    "ffmpeg-python>=0.2.0",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pydantic>=2.11.9",
    "sd-parsers>=0.6",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["pylizlib*"]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
    "ty>=0.0.1a21",
]
